Use packaged js copies when possible.
Exclude the install of jquery.js and underscore.js, and instead
depend and symlink to the packaged versions.

Add sphinx_mathjax_location.patch: point at the local filesystem
path of the mathjax javascript, to avoid a privacy breach when opening
the docs in a browser.

Add depends on libjs-sphinxdoc and libjs-mathjax.